pyproject.toml
[project]
name = "hello"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "mpremote>=1.26.1",
    "pyserial>=3.5",
]
